+++ src/required.js
@@ -1,13 +1,14 @@
-import ValidationResult from './ValidationResult';
+import validator from './validator';
 
 export default function requiredValidator(props) {
     props = Object.assign({}, props);
     props.message = props.message || 'Required';
 
-    // Since validator only performs validation for truthy values,
-    // we do the evaluation directly and return a ValidationResult
-    return (value) => {
-        const isValid = !!value;
-        return new ValidationResult(isValid, props);
-    };
+    // The required validator doesn't ignore any values
+    props.isIgnored = () => false;
+
+    return validator(
+        (value) => !!value,
+        props
+    );
 }